hotfix : 영상렉걸림 해결
