+++ components/layout/Footer.tsx
@@ -2,7 +2,7 @@ import React from "react";
 
 const Footer = () => {
     return (
-        <div className="bottom-0 px-[185px] py-7 flex justify-between">
+        <div className="bottom-0 py-7 flex justify-evenly items-center">
             <div>이용약관</div>
             <div>개인정보처리방침</div>
             <div>사이트맵</div>
