+++ components/layout/Header.tsx
@@ -1,13 +1,11 @@
 import Link from "next/link";
 import React from "react";
-import DropdownMenu from "../category/DropdownMenu";
-import MenuBtn from "../category/MenuBtn";
+import MenuBtn from "../main/category/MenuBtn";
 
 const Header = () => {
     return (
         <div className="flex py-7 justify-between items-center relative">
             <MenuBtn />
-            {/* <DropdownMenu /> */}
             <div className="space-x-3 mr-[185px]">
                 <Link href="/mainPage">메인페이지</Link>
                 <Link href="/communityPage">커뮤니티</Link>
